ATTENTION: Reconfiguring of your source tree is required after this            update!
The mechanism for serialization should now be powerful enough to adapt
the remaining code and implement other kinds of
SERIALIZABLE_ASCII_FILE's.  So it's the best time for any kinds of
reports to this topic, in example to the mailing list.  A simple
Python snippet to feel what's going on [after execution take a look to
the file `blaah.asc' and `blaah2.asc'...  A `diff' of these should be
hopefully an empty string ;) ]:

  *** snip ***

from numpy import *; from numpy.random import randn
from Features import *; from Classifier import *
from Kernel import *; from Library import *
from PreProc import *

num=1000; dist=1;

data=concatenate((randn(2,num)-dist, randn(2,num)+dist), axis=1)
lab=concatenate((-ones(num), ones(num)));

feats=RealFeatures(data);
kernel=GaussianKernel(feats, feats, 2.1);

labels=Labels(lab);

print "*** Begin\n"

svm = GMNPSVM(1, kernel, labels)
feats.add_preproc(NormOne())
feats.add_preproc(NormOne())
feats.add_preproc(NormOne())

feats.print_serializable()

fstream = SerializableAsciiFile("blaah.asc", "w")
if feats.save_serializable(fstream): print "OK writing\n"
else: print "ERROR writing\n"

#svm = GMNPSVM()
feats = RealFeatures()
fstream = SerializableAsciiFile("blaah.asc", "r")
if feats.load_serializable(fstream): print "OK reading\n"
else: print "ERROR reading\n"

fstream = SerializableAsciiFile("blaah2.asc", "w")
feats.save_serializable(fstream)

print "*** End\n"

  *** snap ***

+++ src/libshogun/lib/BitString.h
@@ -81,7 +81,7 @@ class CBitString : public CSGObject
 			uint64_t w=0;
 			int32_t nbits=alphabet->get_num_bits();
 			uint64_t j=0;
-			int32_t nfit=8*sizeof(w)/nbits;
+			uint64_t nfit=8*sizeof(w)/nbits;
 			for (uint64_t i=0; i<len; i++)
 			{
 				w= (w << nbits) | alphabet->remap_to_bin((uint8_t) str[j]);
@@ -158,7 +158,7 @@ class CBitString : public CSGObject
 					string=new uint64_t[stream_len];
 					length=len;
 
-					int32_t idx=0;
+					uint64_t idx=0;
 					int32_t k=0;
 
 					for (int32_t j=0; j<fasta_len; j++, k++)
